[add] latest products uploaded and featured product into home page
