+++ src/Frontend/src/Styles/components/listProducts.css
@@ -2,4 +2,5 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    width: 100%;
   }
