+++ src/Frontend/src/Styles/components/media.css
@@ -1,4 +1,4 @@
-@media only screen and (min-width: 400px){
+@media only screen and (min-width: 300px){
 
     .products-list-container {
         display: grid;
